UE_bot/requirements.txt
notion-client>=2.2.1,<4.0.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
requests>=2.31.0
ddgs>=0.5.0
duckduckgo-search>=6.2.0
